chore(src): extract api crate
