+++ crates/api/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "kickbase"
+name = "api"
 version = "0.1.1"
 edition = "2021"
 
 [dependencies]
-api = { path = "../api" }
-askama = "0.12.1"
-askama_axum = "0.4.0"
 axum = "0.7.5"
 reqwest = { version = "0.12.7", features = ["json"] }
 serde = { version = "1.0.209", features = ["derive"] }
