--- crates/server/src/server/constants.rs
@@ -1,1 +1,0 @@
-pub const API: &str = "https://api.kickbase.com";
